Tag creator refactor (#285)
* Extracts TitleSection on TagCreator

* Adds AffililatedQuestionSection

* Adds AffililatedQuestionSection

* More tag creator refactor

* Adds typography

* Handles codeclimate errors

* Handles simple page load

* Adjusts line length
