+++ cypress/integration/pages/tag-creator.spec.js
@@ -1,0 +1,11 @@
+describe('Tag Creator (Tag Generator)', () => {
+  
+    before(() => {
+      cy.visit('/tag-generator')
+    })
+  
+    it('loads', () => {
+      cy.get('h2')
+        .contains('Tag Generator')
+    })
+  })
